shelf management working with all features

+++ src/laboratory/views/shelfs.py
@@ -16,6 +16,7 @@ from django.urls.base import reverse
 from django.utils.decorators import method_decorator
 from djgentelella.forms.forms import GTForm
 from djgentelella.widgets import core as genwidgets
+from djgentelella.widgets.color import HorizontalBarColorInput, DefaultColorInput
 from weasyprint.text.line_break import get_next_word_boundaries
 
 from laboratory.decorators import has_lab_assigned
@@ -64,13 +65,17 @@ def list_shelf(request, org_pk, lab_pk):
 @has_lab_assigned()
 @permission_required('laboratory.delete_shelf')
 def ShelfDelete(request, lab_pk, pk, row, col, org_pk):
+    if request.method == 'POST':
+        shelf = get_object_or_404(Shelf, pk=pk)
+        shelf.delete()
+        organilab_logentry(request.user, shelf, DELETION, relobj=lab_pk)
+        return {'result': "OK"}
+
     row, col = int(row), int(col)
-    shelf = get_object_or_404(Shelf, pk=pk)
-    shelf.delete()
-    organilab_logentry(request.user, shelf, DELETION, relobj=lab_pk)
-    url = reverse('laboratory:shelf_delete', args=(lab_pk, pk, row, col,org_pk))
+    url = reverse('laboratory:shelf_delete', kwargs={"lab_pk": lab_pk, "pk":pk, "row": row,
+                                                     "col": col,"org_pk": org_pk})
     return {'inner-fragments': {
-        "#modalclose": """<script>$("a[href$='%s']").closest('li').remove();</script>""" % (url)
+        "#modalclose": """<script>delete_shelf("#shelf_%s", "%s");</script>""" % (pk, url)
     },}
 
 
@@ -80,11 +85,12 @@ class ShelfForm(forms.ModelForm,GTForm):
 
     class Meta:
         model = Shelf
-        fields = ['name', 'type', 'furniture']
+        fields = ['name', 'type', 'furniture', 'color']
         widgets = {
             'name': genwidgets.TextInput,
             'type':genwidgets.Select(),
-            'furniture': forms.HiddenInput()
+            'furniture': forms.HiddenInput(),
+            'color':  genwidgets.ColorInput,
         }
 
 
@@ -125,13 +131,13 @@ class ShelfCreate(AJAXMixin, CreateView):
              "ccol": col,
              "data": self.object,
              "org_pk": self.org,
-             "laboratory": self.lab})
+             "laboratory": self.lab}, request=self.request)
         return {
             'inner-fragments': {
                 "#modalclose": "<script>closeModal();</script>"
             },
             'append-fragments': {
-                '#row_%d_col_%d ul' % (row, col): dev,
+                '#row_%d_col_%d ul.sortableself' % (row, col): dev,
             }
         }
 
@@ -169,6 +175,11 @@ class ShelfEdit(AJAXMixin, UpdateView):
         kwargs['initial']['row'] = self.row
         return kwargs
 
+    def get_context_data(self, **kwargs):
+        context=super().get_context_data(**kwargs)
+        context['request'] = self.request
+        return context
+
     def form_valid(self, form):
         self.object = form.save(commit=False)
         furniture = form.cleaned_data['furniture']
@@ -190,7 +201,8 @@ class ShelfEdit(AJAXMixin, UpdateView):
             {"crow": row,
              "ccol": col,
              "data": self.object,
-             "laboratory": self.lab})
+             "org_pk": self.org,
+             "laboratory": self.lab}, request=self.request)
 
         return {
             'inner-fragments': {
